A full fletched development server for greenplum database
It is based on vagrant, ansible and virtualbox and should work on most
host os's yet only tested on mac.

Apart from setting up the server with the basic packages, it also creates
one user with sudo access, ssh login credentials, and global git
configuration in place.

The gpdb repo is not checked out, do this either via git, or scp it or
use a synch folder if you wish. The last option is the slowest for
active development, unless one wishes to cp it to a local folder in the
remote.